environment-class {methods}	R Documentation

Class "environment"

Description

A formal class for R environments.

Objects from the Class

Objects can be created by calls of the form new("environment", ...). The arguments in ..., if any, should be named and will be assigned to the newly created environment.

Methods

coerce

signature(from = "ANY", to = "environment"): calls as.environment.

initialize

signature(object = "environment"): Implements the assignments in the new environment. Note that the object argument is ignored; a new environment is always created, since environments are not protected by copying.

See Also

new.env

[Package methods version 3.2.2 Index]