LinearMethodsList-class {methods}	R Documentation

Class "LinearMethodsList"

Description

A version of methods lists that has been “linearized” for producing summary information. The actual objects from class "MethodsList" used for method dispatch are defined recursively over the arguments involved.

Objects from the Class

The function linearizeMlist converts an ordinary methods list object into the linearized form.

Slots

methods:

Object of class "list", the method definitions.

arguments:

Object of class "list", the corresponding formal arguments.

classes:

Object of class "list", the corresponding classes in the signatures.

fromClasses:

Object of class "list"

Future Note

The current version of linearizeMlist does not take advantage of the MethodDefinition class, and therefore does more work for less effect than it could. In particular, we may move to redefine both the function and the class to take advantage of the stored signatures. Don't write code depending precisely on the present form, although all the current information will be obtainable in the future.

See Also

Function linearizeMlist for the computation, and MethodsList-class for the original, recursive form.

[Package methods version 2.1.0 Index]