MethodSupport {methods}	R Documentation

Additional (Support) Functions for Methods

Description

These are support routines for computations on formal methods.

Usage

getMethodsForDispatch(f, fdef)

cacheMethod(f, sig, def, args, fdef)

resetGeneric(f, fdef, mlist, where, deflt)

Summary of Functions

resetGeneric:

reset the currently defined methods for this generic to the currently visible methods, looking from environment where. Returns TRUE or FALSE according to whether information for the function was found in the metadata.
Normally not called directly, since changes to methods, attaching and detaching packages all generate a call automatically.

cacheMethod:

Store the definition for this function and signature in the method metadata for the function. Used to store extensions of coerce methods found through inheritance.
No persistent effect, since the method metadata is session-scope only.

getMethodsForDispatch:

Get the current methods list object representing the methods for function f, merged from the various packages and with any additional caching information stored in the allMethods slot.
If methods have not yet been merged, calling getMethodsForDispatch will cause the merge to take place.

[Package methods version 2.1.0 Index]