MethodWithNext-class {methods}	R Documentation

Class MethodWithNext

Description

Class of method definitions set up for callNextMethod

Objects from the Class

Objects from this class are generated as a side-effect of calls to callNextMethod.

Slots

.Data:

Object of class "function"; the actual function definition.

nextMethod:

Object of class "PossibleMethod" the method to use in response to a callNextMethod() call.

excluded:

Object of class "list"; one or more signatures excluded in finding the next method.

target:

Object of class "signature", from class "MethodDefinition"

defined:

Object of class "signature", from class "MethodDefinition"

generic:

Object of class "character"; the function for which the method was created.

Extends

Class "MethodDefinition", directly.
Class "function", from data part.
Class "PossibleMethod", by class "MethodDefinition".
Class "OptionalMethods", by class "MethodDefinition".

Methods

findNextMethod

signature(method = "MethodWithNext"): used internally by method dispatch.

loadMethod

signature(method = "MethodWithNext"): used internally by method dispatch.

show

signature(object = "MethodWithNext")

See Also

callNextMethod, and class MethodDefinition.

[Package methods version 3.6.1 Index]