MethodsList-class {methods}	R Documentation

Class MethodsList, Representation of Methods for a Generic Function

Description

Objects from this class are generated and revised by the definition of methods for a generic function.

Slots

argument:

Object of class "name". The name of the argument being used for dispatch at this level.

methods:

A named list of the methods (and method lists) defined explicitly for this argument, with the names being the classes for which the methods have been defined.

allMethods:

A named list, which may be empty if this object has not been used in dispatch yet. Otherwise, it contains all the directly defined methods from the methods slot, plus any inherited methods.

Extends

Class "OptionalMethods", directly.

[Package methods version 2.1.0 Index]