Session {methods}	R Documentation

Deprecated: Session Data and Debugging Tools

Description

The functions traceOn and traceOff have been replaced by extended versions of the functions trace and untrace, and should not be used.

Usage

sessionData()

traceOn(what, tracer=browseAll, exit=NULL)

traceOff(what)

browseAll()

Details

sessionData:

return the index of the session data in the search list, attaching it if it is not attached.

traceOn:

initialize tracing on calls to function what. The function or expression tracer is called on entry, and the function or expression exit on exit.

traceOff:

turn off tracing of this function.

browseAll:

browse the current stack of function calls.
Uses the function debugger to set up browser calls on the frames. On exit from that function, computation continues after the call to browseAll. Computations done in the frames will have no effect.

References

Chambers, John M. (2008) Software for Data Analysis: Programming with R Springer. (For the R version.)

Chambers, John M. (1998) Programming with Data Springer (For the original S4 version.)

[Package methods version 2.9.2 Index]