methods-deprecated {methods}	R Documentation

Deprecated Functions in Package methods

Description

These functions are provided for compatibility with older versions of R only, and may be defunct as soon as the next release.

Details

The original help page for these functions is often available at help("oldName-deprecated") (note the quotes). Functions in packages other than the methods package are listed in help("pkg-deprecated").

See Also

Deprecated, methods-defunct

[Package methods version 3.6.1 Index]