oldGet {methods}	R Documentation

Old functions to access slots in a class definition

Description

These functions are deprecated.

They do nothing but access a slot in a class definition, and don't even do this consistently with the right name (they date back to the early implementation of the methods package). Higher-level functions for the useful operations (e.g., extends for getExtends) should be used instead.

Usage

getAccess(ClassDef)

getClassName(ClassDef)

getClassPackage(ClassDef)

getExtends(ClassDef)

getProperties(ClassDef)

getPrototype(ClassDef)

getSubclasses(ClassDef)

getVirtual(ClassDef)

Arguments

ClassDef 	 the class definition object

Details

The functions should be replaced by direct access to the slots, or by use of higher-level alternatives.

The functions and corresponding slots are:

getAccess 	 "access"

getClassName 	 "className"

getClassPackage 	 "package"

getExtends 	 "contains"

getProperties 	 "slots"

getPrototype 	 "prototype"

getSubclasses 	 "subclasses"

getVirtual 	 "virtual"

See Also

Class classRepresentation.

[Package methods version 2.7.2 Index]