removeMethod {methods}	R Documentation

Remove a Method

Description

Remove the method for a given function and signature. Obsolete for ordinary applications: Method definitions in a package should never need to remove methods and it's very bad practice to remove methods that were defined in other packages.

Usage


removeMethod(f, signature, where)

Arguments

f, signature, where
As for setMethod().

Value

TRUE if a method was found to be removed.

References

Chambers, John M. (2016) Extending R, Chapman & Hall. (Chapters 9 and 10.)

[Package methods version 3.6.1 Index]