makeClassRepresentation {methods}	R Documentation

Create a Class Definition

Description

Constructs an object of class classRepresentation to describe a particular class. Mostly a utility function, but you can call it to create a class definition without assigning it, as setClass would do.

Usage


makeClassRepresentation(name, slots=list(), superClasses=character(),
                        prototype=NULL, package, validity, access,
                        version, sealed, virtual=NA, where)

Arguments

name
character string name for the class

slots
named list of slot classes as would be supplied to setClass, but without the unnamed arguments for superClasses if any.

superClasses
what classes does this class extend

prototype
an object providing the default data for the class, e.g., the result of a call to prototype.

package
The character string name for the package in which the class will be stored; see getPackageName.

validity
Optional validity method. See validObject, and the discussion of validity methods in the reference.

access
Access information. Not currently used.

version
Optional version key for version control. Currently generated, but not used.

sealed
Is the class sealed? See setClass.

virtual
Is this known to be a virtual class?

where
The environment from which to look for class definitions needed (e.g., for slots or superclasses). See the discussion of this argument under GenericFunctions.

References

Chambers, John M. (2008) Software for Data Analysis: Programming with R Springer. (For the R version.)

Chambers, John M. (1998) Programming with Data Springer (For the original S4 version.)

See Also

setClass

[Package methods version 3.6.1 Index]