substituteDirect {methods}	R Documentation

SubstituteDirect

Description

Substitute for the variables named in the second argument the corresponding objects, substituting into object. The argument frame is a named list; if omitted, the environment of the caller is used.

This function differs from the ordinary substitute in that it treats its first argument in the standard S way, by evaluating it. In contrast, substitute does not evaluate its first argument.

The goal is to replace this with an eval= argument to substitute.

Usage


substituteDirect(object, frame, cleanFunction=TRUE)

[Package methods version 3.6.1 Index]