EmptyMethodsList-class {methods}	R Documentation

Internal Class representing Empty Methods List

Description

Objects from class "EmptyMethodsList" are generated during method selection to indicate failed search (forcing backtracking). Other classes described here are used internally in method dispatch. All these are for internal use.

Usage

## class described below
"EmptyMethodsList"

### Other, virtual classes used in method dispatch
"OptionalMethods"
"PossibleMethod"

Slots

argument:

Object of class "name" the argument names being selected on.

sublist:

Object of class "list" (unused, and perhaps to be dropped in a later version.)

Methods

No methods defined with class "EmptyMethodsList" in the signature.

See Also

Function MethodsListSelect uses the objects; see the class MethodsList for the non-empty methods list objects.

[Package methods version 2.9.2 Index]