MethodSupport {methods}	R Documentation

Additional (Support) Functions for Methods

Description

These are internal support routines for computations on formal methods.

Usage


listFromMethods(generic, where, table)

getMethodsForDispatch(fdef, inherited = FALSE)

cacheMethod(f, sig, def, args, fdef, inherited = FALSE)

resetGeneric(f, fdef, mlist, where, deflt)

Summary of Functions

listFromMethods:

A list object describing the methods for the function generic, supplied either as the function or the name of the function. For user code, the function findMethods or findMethodSignatures is recommended instead, returning a simple list of methods or a character matrix of the signatures.

If where is supplied, this should be an environment or search list position from which a table of methods for the generic will be taken. If table is supplied, this is itself assumed to be such a table. If neither argument is supplied, the table is taken directly from the generic function (that is, the current set of methods defined for this generic).

Returns an object of class "LinearMethodsList" (see LinearMethodsList) describing all the methods in the relevant table.

resetGeneric:

reset the currently defined methods for the generic function named f, found in environment where or explicitly supplied as an argument. Other arguments are obsolete and ignored.

Called for its side effect of resetting all inherited methods in the generic function's internal table. Normally not called directly, since changes to methods and the loading and detaching of packages all generate a call automatically.

cacheMethod:

Store the definition for this function and signature in the method metadata for the function. Used to store extensions of coerce methods found through inheritance, and to cache methods with callNextMethod information.

No persistent effect, since the method metadata is session-scope only.

getMethodsForDispatch:

Get the table of methods (an environment since R version 2.6.0) representing the methods for function f.

For user code, the function findMethods or findMethodSignatures is recommended instead, returning a simple list of methods or a character matrix of the signatures.

[Package methods version 3.6.1 Index]