methodUtilities {methods}	R Documentation

Utility Functions for Methods and S-Plus Compatibility

Description

These are utilities, currently in the methods package, that either provide some functionality needed by the package (e.g., element matching by name), or add compatibility with S-Plus, or both.

Usage

functionBody(fun=sys.function(sys.parent()))

allNames(x)

getFunction(name, generic=TRUE, mustFind=TRUE, where)

el(object, where)

elNamed(x, name, mustFind=FALSE)

formalArgs(def)

Quote()

showDefault(object, oldMethods = TRUE)

initMethodDispatch(where = topenv(parent.frame()))

methodSignatureMatrix(object, sigSlots = c("target", "defined"))

Summary of Functions

allNames:

the character vector of names (unlike names(), never returns NULL).

getFunction:

find the object as a function.

elNamed:

get the element of the vector corresponding to name. Unlike the [, [[, and $ operators, this function requires name to match the element name exactly (no partial matching).

formalArgs:

Returns the names of the formal arguments of this function.

existsFunction:

Is there a function of this name? If generic is FALSE, generic functions are not counted.

findFunction:

return all the indices of the search list on which a function definition for name exists.
If generic is FALSE, ignore generic functions.

showDefault:

Utility, used to enable show methods to be called by the automatic printing (via print.default).
Argument oldMethods controls whether old-style print methods are used for this object. It is TRUE by default if called directly, but FALSE when called from the methods package for automatic printing (to avoid potential recursion).

initMethodDispatch:

Turn on the internal method dispatch code. Called on attaching the package. Also, if dispatch has been turned off (by calling .isMethodsDispatchOn(FALSE)—a very gutsy thing to do), calling this function should turn dispatch back on again.

methodSignatureMatrix:

Returns a matrix with the contents of the specified slots as rows. The slots should be named character strings of the same length. Basically used to turn the signatures of a "MethodDefinition" object into a matrix for printing.

[Package methods version 2.9.2 Index]