methods-defunct {methods}	R Documentation

Defunct Functions in the Methods Package

Description

Defunct functions in package methods.

Usage

getAccess(ClassDef)
getClassName(ClassDef)
getClassPackage(ClassDef)
getExtends(ClassDef)
getProperties(ClassDef)
getPrototype(ClassDef)
getSubclasses(ClassDef)
getVirtual(ClassDef)

getAllMethods(f, fdef, where)
mlistMetaName (name, package)
removeMethodsObject(f, where)

seemsS4Object(object)

allGenerics()

Details

The first block of functions should be replaced by direct access to the slots, or by use of higher-level alternatives.

The functions and corresponding slots are:

getAccess 	 "access"

getClassName 	 "className"

getClassPackage 	 "package"

getExtends 	 "contains"

getProperties 	 "slots"

getPrototype 	 "prototype"

getSubclasses 	 "subclasses"

getVirtual 	 "virtual"

getAllMethods replicated the behavior of getMethods, which should be called instead to obtain approximately the old behavior of getAllMethods. However, a better substitution is to use findMethods, which returns a simpler structure.

mlistMetaName was a utility to return the mangled name for the object that stored the MethodsList representation of the methods for a particular generic function, typically in the environment of a package. These objects are no longer used.

removeMethodsObject removed the metadata object containing methods for f.

seemsS4Object was similar to isS4 and superseded by the introduction of the S4 bit in R 2.4.0.

allGenerics is superseded by getGenerics.

See Also

Defunct

[Package methods version 2.9.2 Index]