methods-package {methods}	R Documentation

Formal Methods and Classes

Description

Formally defined methods and classes for R objects, plus other programming tools, as described in the references.

Details

This package provides the ‘S4’ or ‘S version 4’ approach to methods and classes in a functional language.

See the documentation entries Classes, Methods, and GenericFunctions for general discussion of these topics, at a fairly technical level. Links from those pages, and the documentation of setClass and setMethod cover the main programming tools needed.

For a complete list of functions and classes, use library(help="methods").

Author(s)

R Development Core Team

Maintainer: R Core Team R-core@r-project.org

References

Chambers, John M. (2008) Software for Data Analysis: Programming with R Springer. (For the R version.)

Chambers, John M. (1998) Programming with Data Springer (For the original S4 version.)

[Package methods version 2.9.2 Index]